refactor and clean up some stuff
